+++ tests/test_uci.cpp
@@ -47,3 +47,78 @@ TEST(TimeManager, BlackUsesBlackClock) {
     auto alloc = uci::allocate_time(tc, BLACK, 500);
     EXPECT_LT(alloc.iterations, 500);
 }
+
+// --- UCI Protocol Handler Tests ---
+
+#include "uci/uci.h"
+#include "core/attacks.h"
+
+class UCITest : public ::testing::Test {
+protected:
+    static void SetUpTestSuite() { attacks::init(); }
+};
+
+TEST_F(UCITest, UciCommandReturnsUciok) {
+    std::istringstream in("uci\nquit\n");
+    std::ostringstream out;
+    mcts::RandomEvaluator eval;
+    uci::UCIHandler handler(eval, {}, in, out);
+    handler.loop();
+    EXPECT_NE(out.str().find("id name ChessAI"), std::string::npos);
+    EXPECT_NE(out.str().find("uciok"), std::string::npos);
+}
+
+TEST_F(UCITest, IsReadyReturnsReadyok) {
+    std::istringstream in("isready\nquit\n");
+    std::ostringstream out;
+    mcts::RandomEvaluator eval;
+    uci::UCIHandler handler(eval, {}, in, out);
+    handler.loop();
+    EXPECT_NE(out.str().find("readyok"), std::string::npos);
+}
+
+TEST_F(UCITest, PositionStartposAndGo) {
+    std::istringstream in("position startpos\ngo nodes 50\nquit\n");
+    std::ostringstream out;
+    mcts::RandomEvaluator eval;
+    mcts::SearchParams params;
+    params.add_noise = false;
+    uci::UCIHandler handler(eval, params, in, out);
+    handler.loop();
+    EXPECT_NE(out.str().find("bestmove"), std::string::npos);
+}
+
+TEST_F(UCITest, PositionStartposMoves) {
+    std::istringstream in("position startpos moves e2e4 e7e5\ngo nodes 50\nquit\n");
+    std::ostringstream out;
+    mcts::RandomEvaluator eval;
+    mcts::SearchParams params;
+    params.add_noise = false;
+    uci::UCIHandler handler(eval, params, in, out);
+    handler.loop();
+    EXPECT_NE(out.str().find("bestmove"), std::string::npos);
+}
+
+TEST_F(UCITest, PositionFen) {
+    std::istringstream in(
+        "position fen rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq e3 0 1\n"
+        "go nodes 50\nquit\n");
+    std::ostringstream out;
+    mcts::RandomEvaluator eval;
+    mcts::SearchParams params;
+    params.add_noise = false;
+    uci::UCIHandler handler(eval, params, in, out);
+    handler.loop();
+    EXPECT_NE(out.str().find("bestmove"), std::string::npos);
+}
+
+TEST_F(UCITest, GoInfiniteAndStop) {
+    std::istringstream in("position startpos\ngo infinite\nstop\nquit\n");
+    std::ostringstream out;
+    mcts::RandomEvaluator eval;
+    mcts::SearchParams params;
+    params.add_noise = false;
+    uci::UCIHandler handler(eval, params, in, out);
+    handler.loop();
+    EXPECT_NE(out.str().find("bestmove"), std::string::npos);
+}